+++ scripts/macros/generic/noConfiguration.js
@@ -1,0 +1,3 @@
+export async function noConfiguration({speaker, actor, token, character, item, args, scope, workflow}) {
+    workflow.options.configureDialog = false;
+}